Rename usb_console to ptest (#109)
