+++ .travis.yml
@@ -14,7 +14,7 @@ matrix:
         - pip install -r requirements.txt
       script:
         - python -m pytest
-        #- python -m usb_console.run_simulation
+        #- python -m ptest.run_simulation
         - cd ..
         - platformio test --verbose --environment ci_github
         - platformio run --environment ci_github
